add User Profile Icon
