+++ user-react/src/App.js
@@ -9,6 +9,7 @@ import Notifications from "./components/notifications/NotificationsBody";
 import Explore from "./components/explore/explore";
 import Message from "./components/Message/Message";
 import Profile from "./components/profile/profile";
+import UserSignButton from './components/usersignButton/userSignButton';
 
 function App() {
     return (
@@ -28,6 +29,7 @@ function App() {
                             <Route path={"notifications"} element={<Notifications/>}/>
                             <Route path={"Message"} element={<Message/>}/>
                             <Route path={"Profile"} element={<Profile/>}/>
+                            <Route path={"UserSignButton"} element={<UserSignButton/>}/>
                         </Routes>
                     </main>
                     <aside className="col-4 d-none d-md-flex flex-grow-0 p-0">
