+++ user-react/src/components/TrendsOpenion/trendsOpenion.js
@@ -8,15 +8,21 @@ const TrendsOpenion = (props) => {
             <i class="fa-solid fa-ellipsis fa-1x dropdown-toggl text-dark p-1 me-3"
                    data-bs-toggle="dropdown" aria-expanded="false">
 
-                    <ul class="dropdown-menu trends">
+                    <ul class="dropdown-menu trends" style={{fontSize:10,}}>
                         {/* <!-- Dropdown menu links --> */}
-                        <li class=" px-4 py-1 " style={{fontSize:9}}><i class="fa-regular fa-face-frown mx-2"></i>
-                            {props.interestedTrend}
-                        </li>
-                        <hr/>
-                        <li class=" px-4 py-1 " style={{fontSize:9}}><i class="fa-regular fa-face-frown mx-2"></i>
-                            {props.harmfulTrend}
-                        </li>
+                        <div className='d-flex justify-content-center align-items-center' style={{height:'30px'}}>
+                            <li class="  py-1 " style={{fontSize:9}}><i class="fa-regular fa-face-frown mx-2"></i>
+                            
+                            </li>
+                                                <span className=''> {props.interestedTrend}</span>
+                        </div>
+                        <hr style={{margin:0}}/>
+                        <div className='d-flex justify-content-center align-items-center' style={{height:'30px'}} >
+                            <li class=" py-1 " style={{fontSize:9}}><i class="fa-regular fa-face-frown mx-2"></i>
+                            
+                            </li>
+                            <span> {props.harmfulTrend}</span>
+                        </div>
                     </ul>
                 </i> 
         
@@ -24,3 +30,31 @@ const TrendsOpenion = (props) => {
 }
 
 export default TrendsOpenion;
+
+export const UserSignDropdown=(props)=>{
+    return (<>
+    
+    <i class="fa-solid fa-ellipsis fa-1x dropdown-toggl text-dark p-1 me-3"
+                   data-bs-toggle="dropdown" aria-expanded="false">
+
+                    <ul class="dropdown-menu trends" style={{fontSize:10,}}>
+                        {/* <!-- Dropdown menu links --> */}
+                        <div className='d-flex justify-content-center align-items-center' style={{height:'30px'}}>
+                          
+                                                <span className=''> {props.addAccount}</span>
+                        </div>
+                        <hr style={{margin:0}}/>
+                        <div className='d-flex justify-content-center align-items-center' style={{height:'30px'}} >
+                          
+                            <span> {props.logOut}</span>
+                        </div>
+                    </ul>
+                </i> 
+
+    </> );
+    }
+    
+
+
+
+
